+++ src/server.ts
@@ -1,6 +1,6 @@
 import { serve } from "@hono/node-server";
-import { app } from "@/app.ts";
-import { env } from "@/env.ts";
+import { app } from "@/app";
+import { env } from "@/env";
 
 serve(
   {
